requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
appdirs==1.4.4            # via -r requirements.in
argon2-cffi==20.1.0       # via -r requirements.in
asgiref==3.2.7            # via -r requirements.in, django
astroid==2.4.1            # via -r requirements.in
attrs==19.3.0             # via -r requirements.in
bcrypt==3.1.7             # via -r requirements.in
beautifulsoup4==4.9.0     # via -r requirements.in, django-bootstrap4
certifi==2020.4.5.1       # via -r requirements.in, requests
cffi==1.14.0              # via -r requirements.in, argon2-cffi, bcrypt
chardet==3.0.4            # via -r requirements.in, requests
click==7.1.2              # via -r requirements.in
colorama==0.4.3           # via -r requirements.in
defusedxml==0.6.0         # via -r requirements.in, python3-openid
django-allauth==0.41.0    # via -r requirements.in
django-bootstrap4==1.1.1  # via -r requirements.in
django-guardian==2.2.0    # via -r requirements.in
django==3.0.6             # via -r requirements.in, django-allauth, django-guardian
gitdb2==4.0.2             # via -r requirements.in
gitdb==4.0.5              # via gitdb2, gitpython
gitpython==3.1.2          # via -r requirements.in
idna==2.9                 # via -r requirements.in, requests
isort==4.3.21             # via -r requirements.in
lazy-object-proxy==1.4.3  # via -r requirements.in, astroid
mccabe==0.6.1             # via -r requirements.in
oauthlib==3.1.0           # via -r requirements.in, requests-oauthlib
pathspec==0.8.0           # via -r requirements.in
pbr==5.4.5                # via -r requirements.in, stevedore
pillow==7.1.2             # via -r requirements.in
psycopg2==2.8.5           # via -r requirements.in
pycparser==2.20           # via -r requirements.in, cffi
pyparsing==2.4.7          # via -r requirements.in
python3-openid==3.1.0     # via -r requirements.in, django-allauth
pytz==2020.1              # via -r requirements.in, django
pyyaml==5.3.1             # via -r requirements.in
regex==2020.5.14          # via -r requirements.in
requests-oauthlib==1.3.0  # via -r requirements.in, django-allauth
requests==2.23.0          # via -r requirements.in, django-allauth, requests-oauthlib
six==1.14.0               # via -r requirements.in, argon2-cffi, astroid, bcrypt, stevedore
smmap2==3.0.1             # via -r requirements.in
smmap==3.0.4              # via gitdb, smmap2
snowballstemmer==2.0.0    # via -r requirements.in
soupsieve==2.0            # via -r requirements.in, beautifulsoup4
sqlparse==0.3.1           # via -r requirements.in, django
stevedore==1.32.0         # via -r requirements.in
toml==0.10.1              # via -r requirements.in
typed-ast==1.4.1          # via -r requirements.in
urllib3==1.25.9           # via -r requirements.in, requests
wrapt==1.12.1             # via -r requirements.in, astroid
